Fixup FFI

+++ ext/ffi-examples/calc/src/lib.rs
@@ -1,4 +1,4 @@
-// {{{ We'll look at this later...
+// {{{ Just imports up here...
 #![allow(dead_code)]
 
 extern crate libc;
@@ -6,17 +6,17 @@ extern crate libc;
 use libc::c_char;
 use std::ffi::CStr;
 use std::str;
+// }}}
 
-// {{{ fn calculate(c_buf: *const c_car) -> i64 {...}
+// {{{ fn calculate(c_buf: *const c_char) -> i32 {...}
 #[no_mangle]
-pub extern fn calculate(c_buf: *const c_char) -> i64 {
+pub extern fn calculate(c_buf: *const c_char) -> i32 {
     let buf = unsafe { CStr::from_ptr(c_buf).to_bytes() };
     let slice = str::from_utf8(buf).unwrap();
     calc(slice)
 } // }}}
-// }}}
 
-fn calc(script: &str) -> i64 {
+fn calc(script: &str) -> i32 {
     let mut accumulator = 0;
 
     for c in script.chars() {
